requirements.txt
Flask==3.0.2
Werkzeug==3.0.1
filetype==1.2.0
PyPDF2==3.0.1
python-docx==1.1.0
pytest==8.0.2
nltk==3.8.1
langdetect==1.0.9
scikit-learn==1.4.0
sentence-transformers==2.5.1
python-Levenshtein==0.25.0
requests==2.31.0
beautifulsoup4==4.12.3
scholarly==1.7.11
aiohttp==3.9.3
redis==5.0.1
ratelimit==2.2.1
matplotlib==3.8.2
pdfkit==1.0.0
Jinja2==3.1.3 